conductor/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.11.0=py36h5fab9bb_0
aiohttp=3.7.3=py36h8f6f2f9_2
amqp=2.6.1=pyh9f0ad1d_0
apispec=4.3.0=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
billiard=3.6.3.0=py36h8f6f2f9_3
boto=2.49.0=py_0
boto3=1.17.14=pyhd8ed1ab_0
botocore=1.20.14=pyhd8ed1ab_0
brotlipy=0.7.0=py36h8f6f2f9_1001
bz2file=0.98=py_0
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.1=py_0
cachetools=4.2.1=pyhd8ed1ab_0
celery=4.4.6=pyh9f0ad1d_0
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.14.5=py36hc120d54_0
chardet=3.0.4=py36h9880bd3_1008
click=7.1.2=pyh9f0ad1d_0
click-didyoumean=0.0.3=pyh8c360ce_0
click-repl=0.1.6=py_0
cryptography=3.4.4=py36hc39840e_0
decorator=4.4.2=py_0
falcon=2.0.0=py36h8c4c3a4_3
fastdtw=0.3.4=py36he448a4c_0
flake8=3.8.4=py_0
flower=0.9.5=py36h9f0ad1d_0
future=0.18.2=py36h5fab9bb_3
gast=0.4.0=pyh9f0ad1d_0
gensim=3.8.3=py36h831f99a_0
google-api-core=1.25.1=pyh44b312d_0
google-auth=1.24.0=pyhd3deb0d_0
google-cloud-core=1.5.0=pyhd3deb0d_0
google-cloud-storage=1.35.0=pyhd3deb0d_0
google-crc32c=1.1.2=py36h0208b43_0
google-pasta=0.2.0=pyh8c360ce_0
google-resumable-media=1.2.0=pyhd3deb0d_0
googleapis-common-protos=1.52.0=py36h5fab9bb_1
grpcio=1.35.0=py36h8e87921_0
h5py=3.1.0=nompi_py36hc1bc4f5_100
hdf5=1.10.6=nompi_h7c3c948_1111
humanize=3.2.0=pyhd8ed1ab_0
icu=67.1=he1b5a44_0
idna=2.10=pyh9f0ad1d_0
idna_ssl=1.1.0=py36h9f0ad1d_1001
importlib-metadata=3.4.0=py36h5fab9bb_0
importlib_metadata=3.4.0=hd8ed1ab_0
iniconfig=1.1.1=pyh9f0ad1d_0
jansson=2.13.1=h7f98852_0
jmespath=0.10.0=pyh9f0ad1d_0
joblib=1.0.1=pyhd8ed1ab_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kombu=4.6.11=py36h9f0ad1d_0
krb5=1.17.2=h926e7f8_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libcrc32c=1.1.1=h9c3ff4c_2
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libgomp=9.3.0=h2828fa1_18
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
libnghttp2=1.43.0=h812cca2_0
libopenblas=0.3.12=pthreads_hb3c22a3_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.15.1=h780b84a_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h6de172a_18
libxml2=2.9.10=h68273f3_2
logfmt=0.4=pyh9f0ad1d_0
markdown=3.3.3=pyh9f0ad1d_0
mccabe=0.6.1=py_1
more-itertools=8.7.0=pyhd8ed1ab_0
msgpack-python=1.0.1=py36hff7bd54_0
multidict=5.1.0=py36h8f6f2f9_1
ncurses=6.2=h58526e2_4
networkx=2.5=py_0
numpy=1.19.5=py36h2aa4a07_1
openssl=1.1.1j=h7f98852_0
packaging=20.9=pyh44b312d_0
pandas=1.1.4=py36hd87012b_0
pcre=8.44=he1b5a44_0
pip=20.3.3=py36h06a4308_0
pluggy=0.13.1=py36h5fab9bb_4
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.16=pyha770c72_0
prompt_toolkit=3.0.16=hd8ed1ab_0
protobuf=3.15.1=py36hc4f0c31_0
py=1.10.0=pyhd3deb0d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=pyh9f0ad1d_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.2.0=pyh9f0ad1d_0
pyjwt=1.7.1=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py36h5fab9bb_3
pytest=6.2.2=py36h06a4308_2
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-fastjsonschema=2.15.0=pyhd8ed1ab_0
python_abi=3.6=1_cp36m
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.3.1=py36he6145b8_1
readline=8.0=he28a2e2_2
redis-py=3.5.3=pyh9f0ad1d_0
requests=2.25.1=pyhd3deb0d_0
requests-toolbelt=0.9.1=py_0
rsa=4.7.1=pyh44b312d_0
s3transfer=0.3.4=pyhd8ed1ab_0
scikit-learn=0.23.2=py36hb6e6923_3
scipy=1.5.3=py36h976291a_0
setuptools=49.6.0=py36h5fab9bb_3
six=1.15.0=pyh9f0ad1d_0
smart_open=2.1.0=pyh9f0ad1d_0
sqlalchemy=1.3.23=py36h8f6f2f9_0
sqlite=3.34.0=h74cdb3f_0
tensorboard=1.14.0=py36_0
tensorflow=1.14.0=hc3e5e64_0
tensorflow-base=1.14.0=py36hc3e5e64_0
tensorflow-estimator=1.14.0=py36h5ca1d4c_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
toml=0.10.2=pyhd8ed1ab_0
tornado=6.1=py36h8f6f2f9_1
tqdm=4.56.0=pyhd8ed1ab_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
ujson=4.0.2=py36h2531618_0
urllib3=1.26.3=pyhd8ed1ab_0
uwsgi=2.0.19.1=py36h2e0b79c_1
vine=1.3.0=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.12.1=py36h8f6f2f9_3
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.5.1=py36h8c4c3a4_0
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
